Exposed per-core timer interrupts on SPARC
timer_interrupt() will be called on a timer interrupt.  set_timer(usec) arms
a core's timer.  Beware that, internally, we only support a power-of-2 number
of cycles as the timer period, so the usec value will be rounded up to the
next power-of-2 cycle count.  Also, the max period is 2^24 cycles (~60 HZ for
a 1 GHz target clock).