Trigger the monitor with an interrupt
It will respond to keyboard or serial input (so it works with VMs).  Hit
the char, drop into the monitor.  Exit it to return from the interrupt.
To use it, set the __CONFIG_MONITOR_ON_INT__ variable in your Makelocal.

Careful KFS running from this.  In fact, you don't want to, since it
won't return, meaning you won't return from the keyboard interrupt to
whatever process you were running before.