Made a global num_mgmtcores variable
Replaces the local num_reserved variable in proc_init()
so that it can be used in other functions throughout the file.

Also added a __CONFIG_OSDI__ variable for hacky stuff
related to our upcoming OSDI submission.  Need to think
more about how to fix this properly... The problem is
that the code below this hack is the 'proper' solution,
but our kernel level scheduler currently can't properly
handle processes in the manycore state.