Preemption functions
Half-temporary functions to preempt or warn about preemption for a
process's core (or all cores).  There are also test apps to run while
this is going on (msr_nice_while will yield when warned).  There are
monitor functions added to measure to test these things.

The time to preempt is really low, enough so that I wonder if it's
working or not.