Fixed deadlock issue in pthread barrier wait.
While a thread is waiting on a barrier, it can call pthread_yield and may come
back on a different vcore. Using vcore id as the index into the sense array
is wrong. Use pthread id instead.

In the future, we might dynamically allocate the sense array based on how
many pthreads will wait on a barrier, instead of having the static array.