MPC TLB shootdowns
This should work for now, though it doesn't actually check that other
cores receive the message before returning (which is a slight race).
See the discussion in the documentation for more details.  Could add a
checklist to be safe (though don't put in a barrier, unless you don't
include the caller, which is a minor pain).