FP save/restore on x86
Saves and restores FP regs.  This will also save SSE registers, but
supposedly only if the appropriate bit is set in cr4.  This might not be
logically correct, especially with the _S state, but it will save and
restore them for preemptions, which should be good enough for OSDI.  We
still need to think through the best places to put them and under what
conditions we need to save/restore.

The pthread library still needs to handle it's own save and restore (on
both sparc and x86).

You need to reinstall your kernel headers (or at least x86's
trapframe.h).