Fix dependency on vcoreid in barrier and cond var.
Bthread used vcoreid to find the unique waiter, but that is incorrect in
pthread, since threads can yield and move to another vcore. For the barrier
and condition variable impelmentation, it only requires a unique waiter.
So we allocate that unique waiter from an array of them.

next slot is just an optimization to give us a better chance of finding
an empty slot quickly. So the race on that is benign.