Monitor enhancements
Allows for reading the procinfo (unlocked!) of a process, which would
deadlock the system if the lock was already held.  Also allows for the
unlocking of arbitrary locks (you provide the address).

This will help with killing the EXPER_TRADPROCs, which don't want to die
anymore.  Just unlock them, and they will go away.  Yikes.