Fixed bug in pthread barrier code
Needed to reset some more state after the barrier is cleared, but before
opening the floodgates.  Also needed a memory barrier in the pthread_mutex()
code to ensure the write to the count variable and the write to the lock variable
are in the proper order.  This one was a bitch to track down.