Use correct memory barrier on x86
On Intel implementations of x86, SFENCE seems to only have any effect on
programs that use the write-combining buffers (e.g., those that use the
SSE instruction MOVNTPS).  Its semantics seem to NOT guarantee load-to-store
ordering.

LFENCE and SFENCE do not prevent older stores from bypassing younger
loads to different addresses, which is something a full memory barrier
should most definitely do.  So, mb() should call MFENCE, which does
have this property, not LFENCE+SFENCE.  If the latter two worked, it
may be because they happened to interlock the pipeline just enough that
the desired ordering serendipitously was enforced.