Fixed experiment timing on SPARC
Since target time progresses much more slowly than wall time, I divide
the argument to any udelay call in the microbenchmarks by a constant.
I don't do this in a header file because really we don't want to change
the semantics of udelay, just how long we delay for in these benchmarks.