CPIO support for KFS
Basic support to understand CPIO archives for an initramfs.  There is
some rudimentary support in the Makefile mess to build the initramfs
based on some variables in Makelocal.  If you want to strip the binaries
before adding them, do it in a tools script.  For whatever bullshit
Makefile reason, it's a pain doing that in tests/Makefrag.