mmap() and friends using vm regions
Not heavily tested, esp with files.  There are still things related to
the old pfault_info floating around, which will go away soon.

mprotect() also now is best effort, and won't fail if a part of the
interval is unmapped.  Might change that later.

There are also a bunch of inefficiencies with vmr management, related to
merging and finding them when performing some of these operations.
Plenty of TODOs.