mmap() merges, fixed 0-page allocation bug
do_mmap() will attempt to merge its new VMR, which will help a bit
(although it might not be the best way to do things).

There was also a bug with VMR requests for the 0th page, which is now
fixed.  You can create that vmr, but it's not recommended.  mmap() and
friends will check to make sure userspace doesn't affect the 0th page.
Also note that glibc did not like getting the 0 page, for whatever
reason.