generic_file_write() and file holes
readpage() is supposed to handle hole (use a 0'd page).  KFS can handle
extending files and dealing with the holes, though they won't get saved,
which may be an issue once we start evicting from the page cache.  Also,
there are no constraints on file sizes or anything like that.