KFS defaults (Read this if you have no files)
You don't have to have anything in KFS (default kern/kfs), but you ought
to have a bin/ with your binaries and a lib/ with ld.so.1 and libc.so.6
(and any other libraries you link against).

There's also no automatic copying of test binaries, either to KFS or the
old FS folder, to save dev time, to keep the FS manageable, and to keep
the image size down for people using static apps.