Block requests come with a completion method
breq_submit_request() doesn't require you sleep on the breq, but since
it is common (to both callers!) we provide helpers for it.  This also
splits up the usage of lock_page() to both block on IO as well as block
on waiting for a page to be ready (blocking on someone else's IO).