Changeable kernel stacks
We can now set and query the default kernel stack (the one you'll trap
in to).  On x86, this info is saved in the TSS.  On sparc, it's in an
array.  If that array gets a lot of cache traffic, we'll need to put it
in per_cpu_info, which is a minor pain.