Copies out current_tf to pcpui
This is in response to a pretty bad bug/design issue.  The documentation
has a lot about it and about what to expect in future patches.  The
TODOs and checks in kthread.c will go away in a couple patches.

Ultimately, we want to avoid copying the TF multiple times, but that's
asm for another day.  Recommend writing the location of actual_tf at the
top of the stacktop (avoid computing it, simplify the code, etc).

Note: this compiles on sparc, and might actually work!  Hard to tell,
with that glibc-ish bug floating around.