Syscall return paths cleaned up
The first syscall will return naturally via proc_restartcore() (which
does nothing if the core was already restarted and there is no
current_tf).

This also disables batching for now (not that anyone uses it, or the
asynchrony for that matter).  In the future, we can use kmsgs
(carefully) for batched syscalls, being careful of the *tf passed around
in the kmsg functions...