Support for the Ethernet-Audio device
Turn on support in your Makelocal.  Right now, the kernel copies
whatever it gets into a file (in /dev) that a program (like
test/eth_audio.c) can mmap.  When a packet comes in, the old buffer gets
sent back (including on the first time through).

Also adds some networking infrastructure (checksums, #defines), and the
eth_aud packets are intercepted (in a ghetto manner) from both the e1000
and the rl8186 NICs.