Closes files in proc_destroy()
If there is a syscall in progress on a file, it will stay open til the
end of the call (which is what the kref does).  This change allows us to
have the files (or something they point to) reference the process.
Note, this still doesn't work for mmap() files.  Eagle-eyed readers
should know why!