Eth audio device can notify processes
When you run eth_audio, call it from busybox.  This will keep core 0 in
a loop of BB yielding, which will allow the NIC interrupts to be
received.  When ROS polls in the monitor, it does not receive interrupts
until it reads a complete line.

Note that the kernel cannot handle a process that does not munmap the
VMR before exiting.  I'm not willing to overhaul things for a ghetto
interface/device yet, so just be careful.