Syscalls take event_queues for completion (XCC)
If the syscall has an ev_q, the kernel will try and signal/send an
event.  tests/syscall.c has examples of how userspace can do it (rawly),
for both a polling ev_q as well as for an IPI driven one (which requires
_M mode).

Use SYS_block whenever you want to test a syscall that will block, but
don't use more than one of them at a time or in conjunction with any
blocking disk reads (for now).

Rebuild your cross compiler.