VCORE_APPRO sets the ev_mbox
When the kernel is picking what the vcore ought to be, the kernel will
use the vcore's ev_mbox.  Needs to be this way, since the user doesn't
know which vcore the message is going to (which is the nature of
VCORE_APPRO).  If  want something that allows a message to go to an
ev_q's mbox, but the IPIs to get sent somewhere else, then use something
like Round-Robin.