Resyncs parlib's sparc's atomics with the kernel's
Note that while similar, userland atomics in sparc are slightly
different than their kernel counterparts.  One big difference is that
sparc (and x86 too) doesn't not have an atomic_add().  This name
conflicts with glibc, and is part of a larger problem of having two
sources of atomic primitives that might not line up right.
user/parlib/include needs to work with both ROS and glibc (for now), so
this causes problems.

Another difference is with irqsaves (or lack thereof).

Normally, userspace just needs to get the functions it needs, but it was
behind enough that it just received most of the ones from the kernel.