Fixes memory leak with hashtable_iter
Lots of ughs with this.  Had they not typedef'd, it might have been more
obvious.  The best way would be to have the caller provide the memory
too, so we can use stack vars, but whatever.  3rd party code...

Also changed the remaining two users to use hash_for_each.