Alarm infrastructure
Allows you to block a kthread for a while or to run a handler after a
while.  These alarms are grouped in sorted chains, which are attached to
an interrupt source.  Currently, there are chains for the per-core timer
interrupts.

Check out k/i/alarm.h for a quick howto.