Converts sys_halt_core() to use alarms
Not sure what we want to do long term for core halting.  This will keep
halting the core until the timer is up, servicing interrupts along the
way.  There are a lot of interrupts coming in to core 0 on hardware too.

Note if you want to kfunc sys_halt_core(), your arguments are in hex. =P