Fixes nasty kthread bug
We were freeing the stack we were about to use, which is a bad idea.
Bug showed up as TAILQ corruption (often PFs) in alarm code, and once a
non-empty semaphore in a waiter in sys_block().  The "freed" stack was
being reused for the exact same code path (another sys_block), and
occasionally for other syscalls with a deeper stack (clobbering the LL
item).