queue "SAFE" ops for userspace (XCC)
This is just a more up-to-date sys/queue.h that has the FOREACH_SAFE
ops, allowing you to remove the item from within the for loop.  The
kernel has been using these for a while now.

Rebuild your cross compiler, or at least put the new queue.h in your
XCC headers.