Updated memory barrier stuff
Minor change to the MCS lock (which did not appear to be a problem in
the assembly) and some talk about the Alpha memory model.

Note that I caught a process deadlocked on an MCS lock in a VM (one
vcore, spinning on a locked lock, no preemptions etc outstanding), so
something out there is probably jacked up.